+++ static/posit/specialized/posit_48_2.cpp
@@ -1,6 +1,6 @@
 // posit_48_2.cpp: test suite runner for specialized extended standard 48-bit posit<48,2>
 //
-// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
@@ -52,7 +52,7 @@ try {
 
 	ReportTestSuiteHeader(test_suite, reportTestCases);
 
-	size_t RND_TEST_CASES = 1024;
+	unsigned RND_TEST_CASES = 1024;
 
 	using Scalar = posit<nbits, es>;
 	Scalar p;
